Implemented gradient-filled outlines of rectangles and rounded-rectangles. - On plain old X11 rounded rectangles are drawn using the gradient's first   stop color as a fallback. - With GDI+ rounded corners are asymmetric when the gradient is not   opaque.
GDI+: Fixed asymmetry when filling rounded rectangles with a gradient,
but only when the gradient is opaque since I can't paint the edges more
than once with a transparent gradient.

The rect element -outline option can now be a gradient.

The marquee -outline option can now be a gradient.

X11: Made the dotted-line and dotted-rectangle drawing use XDrawLine instead
of XDrawPoint.