New options -xscrollsmoothing and -yscrollsmoothing fix a long-standing issue where specifying small values for -xscrollincrement or -yscrollincrement would result in 1-pixel-at-a-time scrolling with the scrollbar buttons and when autoscanning.
The default value of -showlines is now false on Mac OS X and Gtk+, true
on Win32 and X11.

The default value of -columnproxy is now "realtime".

Fixed the column drag indicator being displayed in the wrong place when
-canvaspadx was used.

Fixed the loupe command not filling the entire image with pixel data if there
was a partially-visible row or column of zoomed pixels at the right or bottom
edges.  This change could result in the image being resized but only if
the image's -width and -height weren't specified.

New function Tree_GetFlagsFromObj handles all the cases where a string
is parsed into 'x', 'y', 'e, 'w' etc flags.