Disabled caching the -background image for now since it won't work for transparent background images and I don't know of any way to ask Tk if an image has transparency.
Fixed the -backgroundimage not being anchored in the correct position
relative to the bottom/right of the canvas.  Ditto for canvas-relative gradient
coordinates.  This was due to not taking the "fake" whitespace into
account.

Fixed tiling of image elements, the offsets weren't corrected when
Tree_DrawTiledImage was changed for the new bgimage options.

Cache calculation of the canvas fake width and height.

Removed a bunch of conditional-compilation code.