Fixed broken header layout on Mac OS X.
HIThemeGetButtonContentBounds() was returning unexpected results,
so the result of TreeTheme_GetHeaderContentMargins() is now hard-coded
based on the observed width and padding of the sort arrow.

Due to partArrow.padX/Y not being initialized in the column layout code,
the column bitmap/image/text was justified incorrectly/randomly whenever
the sort arrow was displayed.

The space available to the column text was not reduced by the width of
the sort arrow, resulting in the sort arrow being drawn over the text
when the column was narrow enough.