Fixed a bug with vertical style layout and union padding.
The code wrongly assumed that the bottom edge of the bottom-most element was
the farthest to the bottom.  That isn't the case when a preceding element
has a lot of union padding on its bottom edge.  This bug could result in
elements expanding beyond the space available to the style.