Fixed issues with header elements and states.
The header element no longer ignores the STATE_ITEM_XXX and STATE_HEADER_XXX flags.
Previously, only the -state option was used for the per-state options.

COLUMN_STATE_NORMAL was the default if -state was unspecified.  Now the
STATE_HEADER_XXX flags are checked for a suitable value.

COLUMN_STATE_ARROW_NONE was the default if -arrow was unspecified.  Now the
STATE_HEADER_XXX flags are checked for a suitable value.