Rewrote TreeItem_WalkSpans to handle column header drag-and-drop better.
If the programmer specified some illogical drag or indicator columns then
the feedback could look weird.  For example, if a header of span 1 was being
dragged in header1 while a header of span 2 was displayed in header2, the user
would see a span of 2 being dragged even though only 1 column was actually being
moved.

Now the spans of dragged headers are restricted to the number of dragged columns,
and the spans of non-dragged headers are terminated if they reach the first dragged
column.