Improved the column-spanning code.
Previously, a style spanning more than one column would request all of its
width in the first column of the span.  Also, when calculating the needed
height of an item, spans were ignored.

Now, a style distributes its width across all visible columns in the span,
taking into account the -minwidth, -width, and -maxwidth column options.
Also, when calculating the needed height of an item, spans are not ignored.