Improved the column-spanning code again.
Instead of distributing width in columns for every span in every item, just
find the widest span and later use that to distribute width.

Corrected for too much width being given to columns in some cases; try to
keep the total width of columns as small as needed for the styles being
displayed.