Fixed compatibility and column-visibility issues with column spanning.
Compat: When only one column is visible, the width of items can be different
than the width of the column header.

Compat: -indent width is not requested in an item if no TreeItemColumn exists.
It might be good to request -indent width anyway, but a test broke.

Added column-visibility checks to several places.