Fixed broken TrimTheFat() column-spanning code (again).
A previous fix moved TreeColumnPriv.maxSumOfSpanWidths into ColumnReqData
for each column.  There was some problem with that resulting in different
sums of span widths when smaller ranges of columns were updated compared to
when all the columns were updated together.  Possibly it could have been
fixed by updating maxSumOfSpanWidths in columns to the right when a smaller
range of columns was updated.

The new solution is to keep linked lists of adjacent spans from left to
right.  Each ColumnSpan has a list of spans to the right, whereas before it
was only a single pointer to another span in no particular order.  It is
then possible to find the maximum width of each chain of spans for any
range of columns that doesn't overlap another range of columns.